rewrite page_fault_handler()

+++ arch/x86/kernel/include/ihk/types.h
@@ -5,16 +5,16 @@ typedef unsigned char      uint8_t;
 typedef unsigned short     uint16_t;
 typedef unsigned int       uint32_t;
 typedef unsigned long long uint64_t;
-typedef char               int8_t;
-typedef short              int16_t;
-typedef int                int32_t;
-typedef long long          int64_t;
+typedef signed char        int8_t;
+typedef signed short       int16_t;
+typedef signed int         int32_t;
+typedef signed long long   int64_t;
 
-typedef long long          ptrdiff_t;
+typedef int64_t            ptrdiff_t;
 typedef int64_t            intptr_t;
 typedef uint64_t           uintptr_t;
-typedef unsigned long long size_t;
-typedef long long          ssize_t;
+typedef uint64_t           size_t;
+typedef int64_t            ssize_t;
 typedef int64_t            off_t;
 
 #define NULL ((void *)0)
